+++ package.json
@@ -58,7 +58,8 @@
     "src/loader.js",
     "src/math.js",
     "src/time.js",
-    "src/DebugLayer.js"
+    "src/DebugLayer.js",
+    "src/keycodes.js"
   ],
   "dependencies": {
     "grunt": "~0.4.1",
